mediastreamer2 DCCP Support: Bitrate Control, Part 2--Improvements
Update the video encoder to return an average of its recent bitrate instead of simply its requested bitrate to improve the changes that bitrate control can make. Also update increase/decrease bitrate functions to ensure that they are attempting to set bitrates that are vaguely reasonable (within a factor of 10 of recent bitrate).